Tests/LibWeb/Text/expected/fit-content-height-uses-resolved-width.txt

wrapper: 456x260
canvas: 448x252
height delta: 8
